Rename folder and add npm dependencies
