+++ metrics/index.ts
@@ -1,2 +1,2 @@
 export { initializeOtel } from './init';
-export { incrementOtelCounter, observeOtelGauge, recordOtelHistogram } from './instrumentation/instruments';
+export { incrementOtelCounter, observeOtelGauge, recordOtelHistogram } from './src/instruments';
