+++ metrics/package.json
@@ -20,5 +20,13 @@
   "bugs": {
     "url": "https://github.com/boxyhq/jackson/issues"
   },
-  "homepage": "https://github.com/boxyhq/jackson#readme"
+  "homepage": "https://github.com/boxyhq/jackson#readme",
+  "dependencies": {
+    "@opentelemetry/api": "1.4.1",
+    "@opentelemetry/exporter-metrics-otlp-grpc": "0.37.0",
+    "@opentelemetry/exporter-metrics-otlp-http": "0.37.0",
+    "@opentelemetry/resources": "1.11.0",
+    "@opentelemetry/sdk-metrics": "1.11.0",
+    "@opentelemetry/semantic-conventions": "1.11.0"
+  }
 }
