+++ metrics/src/instruments/gauge.ts
